Fix bug causing user-supplied start values to be ignored
model.run() included a named 'start' parameter for passing to the pymc3
sampler, but also had code that assumed 'start' would be part of kwargs,
with the result that user-supplied start values were always overwritten
by default values. This commit simply removes 'start' from the method
signature, so it will always be handled as a kwarg instead.