Fix level_map issue
When adding a fixed term that has already been indirectly referenced in a previous call to add_term (for example, through the split_by or orthogonalize arguments), the level mapping for the fixed term was never getting added to level_map. (But everything worked fine if the fixed term was added BEFORE ever getting referenced in a split_by or orthogonalize argument.) It turns out that this can be fixed by a fairly simple tweak to _get_variable_data, namely, assigning default labels to all 'variable' arguments and hashing labels rather than variables.